requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu


fastapi
uvicorn

# LangChain + LLM
langchain-core
langchain-groq
langchain-text-splitters

# Embeddings
langchain-huggingface
sentence-transformers

# Vector DB
pinecone

# File parsing
pymupdf
python-docx
extract-msg
beautifulsoup4

# Utilities
requests
python-multipart

python-dotenv

